# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NetSwift API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Order {
  id: number;
  orderId: string;
  network: string;
  bundleSize: string;
  bundleValidity: string;
  phone: string;
  status: string;
  gbAmount: number;
  price: number;
  createdAt: string;
}

export interface CreateOrderBody {
  network: string;
  bundleSize: string;
  bundleValidity: string;
  phone: string;
  gbAmount: number;
  price: number;
}

export interface DashboardStats {
  walletBalance: number;
  ordersToday: number;
  gbSoldToday: number;
  successRate: number;
  totalOrders: number;
}

export interface Transaction {
  id: number;
  type: string;
  amount: number;
  description: string;
  /** @nullable */
  reference: string | null;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListOrdersParams = {
  network?: string;
  status?: string;
  limit?: number;
};

export type TrackOrderParams = {
  /**
   * Order ID (e.g. NSW-XXXXXXXX) or phone number
   */
  q: string;
};

export type ListTransactionsParams = {
  limit?: number;
};