# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NetSwift API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all orders
 */
export const listOrdersQueryLimitDefault = 20;

export const ListOrdersQueryParams = zod.object({
  network: zod.coerce.string().optional(),
  status: zod.coerce.string().optional(),
  limit: zod.coerce.number().default(listOrdersQueryLimitDefault),
});

export const ListOrdersResponseItem = zod.object({
  id: zod.number(),
  orderId: zod.string(),
  network: zod.string(),
  bundleSize: zod.string(),
  bundleValidity: zod.string(),
  phone: zod.string(),
  status: zod.string(),
  gbAmount: zod.number(),
  price: zod.number(),
  createdAt: zod.string(),
});
export const ListOrdersResponse = zod.array(ListOrdersResponseItem);

/**
 * @summary Create a new order
 */
export const CreateOrderBody = zod.object({
  network: zod.string(),
  bundleSize: zod.string(),
  bundleValidity: zod.string(),
  phone: zod.string(),
  gbAmount: zod.number(),
  price: zod.number(),
});

/**
 * @summary Track an order by order ID or phone number
 */
export const TrackOrderQueryParams = zod.object({
  q: zod.coerce
    .string()
    .describe("Order ID (e.g. NSW-XXXXXXXX) or phone number"),
});

export const TrackOrderResponse = zod.object({
  id: zod.number(),
  orderId: zod.string(),
  network: zod.string(),
  bundleSize: zod.string(),
  bundleValidity: zod.string(),
  phone: zod.string(),
  status: zod.string(),
  gbAmount: zod.number(),
  price: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Get a single order by ID
 */
export const GetOrderParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOrderResponse = zod.object({
  id: zod.number(),
  orderId: zod.string(),
  network: zod.string(),
  bundleSize: zod.string(),
  bundleValidity: zod.string(),
  phone: zod.string(),
  status: zod.string(),
  gbAmount: zod.number(),
  price: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Get dashboard statistics
 */
export const GetDashboardStatsResponse = zod.object({
  walletBalance: zod.number(),
  ordersToday: zod.number(),
  gbSoldToday: zod.number(),
  successRate: zod.number(),
  totalOrders: zod.number(),
});

/**
 * @summary List all transactions
 */
export const listTransactionsQueryLimitDefault = 20;

export const ListTransactionsQueryParams = zod.object({
  limit: zod.coerce.number().default(listTransactionsQueryLimitDefault),
});

export const ListTransactionsResponseItem = zod.object({
  id: zod.number(),
  type: zod.string(),
  amount: zod.number(),
  description: zod.string(),
  reference: zod.string().nullable(),
  createdAt: zod.string(),
});
export const ListTransactionsResponse = zod.array(ListTransactionsResponseItem);