# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NetSwift API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderBody {
  network: string;
  bundleSize: string;
  bundleValidity: string;
  phone: string;
  gbAmount: number;
  price: number;
}