# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NetSwift API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  walletBalance: number;
  ordersToday: number;
  gbSoldToday: number;
  successRate: number;
  totalOrders: number;
}