# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NetSwift API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  orderId: string;
  network: string;
  bundleSize: string;
  bundleValidity: string;
  phone: string;
  status: string;
  gbAmount: number;
  price: number;
  createdAt: string;
}