# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NetSwift API
 * OpenAPI spec version: 0.1.0
 */

export interface Transaction {
  id: number;
  type: string;
  amount: number;
  description: string;
  /** @nullable */
  reference: string | null;
  createdAt: string;
}